repo: remove check for '.git' at end of git remote
I can't think of a reason to enforce this. When script is run by default
(at least with git version 2.42.0), there is no '.git' at the end of the
remote.

Here is what 'git remote -v' shows:
  origin  https://github.com/netbox-community/devicetype-library

After removing the check, we can run the script multiple times in a row
without deleting the 'repo' folder.

Fixes issue #129